refactor: fix spelling mistakes, format Cargo.toml and remove unwanted files
