--- runtime/lite/src/modules/wasmtime_wasi_module/help.rs
@@ -1,11 +1,0 @@
-// use wiremock::MockServer;
-
-// async fn create_server(add: &str) -> MockServer {
-//     let listener = std::net::TcpListener::bind(add).unwrap();
-//     MockServer::builder().listener(listener).start().await
-// }
-
-// pub async fn post(address: &str) -> MockServer {
-//     let server = create_server(address).await;
-//     server
-// }
