+++ runtime/lite/src/modules/wasmtime_wasi_module/mod.rs
@@ -1,16 +1,13 @@
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use std::sync::MutexGuard;
-pub mod help;
-pub use help::*;
-mod tests;
-
 use crate::modules::state_manager::{
     ExecutionState, GlobalState, GlobalStateManager, WorkflowState,
 };
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
 use sha256::digest;
+use std::sync::MutexGuard;
 use std::sync::{Arc, Mutex};
 
+mod tests;
 mod types;
 pub use types::*;
